Move `WaitingReceiver` to its own module
This component houses a good chunk of functionality now and is
better understood in isolation.

Additionally, this allows us to better enforce visibility of types.
